chore(examples): rename repro-named examples to feature-named examples
Examples should read as real examples, not bug reproductions; the
regression value is unchanged (same kernels, same assertions).

- issue150_repro            -> toolchain_pairing_smoke
- struct_layout_repro       -> struct_field_layout
- addressof_sharedarray_repro -> addressof_sharedarray
- closure_index_miscompile  -> ref_index_projections

Headers rewritten in example voice (the feature each exercises first, a
one-line 'guards the fix for issue #N' note last); package names, lock
files, README table, smoketest LTOIR list, and the mir-lower unit-test
comment all updated.

GPU-verified on sm_120: struct_field_layout, ref_index_projections and
addressof_sharedarray pass; toolchain_pairing_smoke compiles.
