+++ crates/rustc-codegen-cuda/examples/addressof_sharedarray/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "addressof_sharedarray_repro"
+name = "addressof_sharedarray"
 version = "0.1.0"
 edition = "2024"
 authors = ["Nihal Pasham <nihalp@nvidia.com>"]
