+++ crates/rustc-codegen-cuda/examples/ref_index_projections/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "closure_index_miscompile"
+name = "ref_index_projections"
 version = "0.1.0"
 edition = "2024"
 license = "Apache-2.0"
