+++ crates/rustc-codegen-cuda/examples/toolchain_pairing_smoke/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "issue150_repro"
+name = "toolchain_pairing_smoke"
 version = "0.1.0"
 edition = "2024"
 license = "Apache-2.0"
